requirements.txt
Django==2.2.10
django-configurations==2.2
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==6.2.0
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
